src/embeddings-api/requirements.txt
fastapi
dapr-ext-fastapi-dev
uvicorn
requests
environs
typing-extensions
httpx
agnt-smth
sentence-transformers
langchain
langchain-chroma
langchain-community
langchain-huggingface
tiktoken
